+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * handle unhandled rejection which was rejected to an undefined value
 * add more context to the express app requests middleware stack
 * adds logging to all oauth handler steps
+* upgrade hull-client to v1.1.1 to add support for `scopes` claim in auth tokens
 
 # 0.11.9
 * sqs adapter